+++ packages/backend-common/src/tokens/ServerTokenManager.test.ts
@@ -171,7 +171,6 @@ describe('ServerTokenManager', () => {
 
     it('should throw for expired tokens, and re-issue new ones', async () => {
       jest.useFakeTimers();
-      const warn = jest.spyOn(logger, 'warn');
 
       const tokenManager = ServerTokenManager.fromConfig(configWithSecret, {
         logger,
@@ -179,14 +178,12 @@ describe('ServerTokenManager', () => {
 
       const { token: token1 } = await tokenManager.getToken();
       await expect(tokenManager.authenticate(token1)).resolves.not.toThrow();
-      expect(warn.mock.calls.length).toBe(0);
 
       // Right before the reissue timeout, it still returns the same token
       jest.advanceTimersByTime(9 * 60 * 1000);
       const { token: token1Again } = await tokenManager.getToken();
       expect(token1).toEqual(token1Again);
       await expect(tokenManager.authenticate(token1)).resolves.not.toThrow();
-      expect(warn.mock.calls.length).toBe(0);
 
       // Right after the reissue timeout, the old ones are still valid but returning a new token
       jest.advanceTimersByTime(2 * 60 * 1000);
@@ -194,14 +191,47 @@ describe('ServerTokenManager', () => {
       expect(token1).not.toEqual(token2);
       await expect(tokenManager.authenticate(token1)).resolves.not.toThrow();
       await expect(tokenManager.authenticate(token2)).resolves.not.toThrow();
-      expect(warn.mock.calls.length).toBe(0);
 
       // After expiry of the first one, it gets warnings but the newest one is still valid
       jest.advanceTimersByTime(52 * 60 * 1000);
-      await expect(tokenManager.authenticate(token1)).resolves.not.toThrow();
+      await expect(tokenManager.authenticate(token1)).rejects.toThrow(
+        'Invalid server token; caused by JWTExpired: "exp" claim timestamp check failed',
+      );
       await expect(tokenManager.authenticate(token2)).resolves.not.toThrow();
-      expect(warn.mock.calls[0][0]).toMatchInlineSnapshot(
-        '"#### DEPRECATION WARNING: #### Server-to-server token had an expired exp claim, support for this has been deprecated and will result in errors in a future release"',
+    });
+
+    it('should work with a manually crafted JWT', async () => {
+      const secret = 'a1b2c3';
+      const token = await new jose.SignJWT({})
+        .setProtectedHeader({ alg: 'HS256' })
+        .setSubject('backstage-server')
+        .setExpirationTime(Date.now() + 1000 * 60 * 60)
+        .sign(jose.base64url.decode(secret));
+
+      const tokenManager = ServerTokenManager.fromConfig(
+        new ConfigReader({
+          backend: { auth: { keys: [{ secret }] } },
+        }),
+        { logger },
+      );
+      await expect(tokenManager.authenticate(token)).resolves.toBeUndefined();
+    });
+
+    it('should reject tokens without exp claim', async () => {
+      const secret = 'a1b2c3';
+      const token = await new jose.SignJWT({})
+        .setProtectedHeader({ alg: 'HS256' })
+        .setSubject('backstage-server')
+        .sign(jose.base64url.decode(secret));
+
+      const tokenManager = ServerTokenManager.fromConfig(
+        new ConfigReader({
+          backend: { auth: { keys: [{ secret }] } },
+        }),
+        { logger },
+      );
+      await expect(tokenManager.authenticate(token)).rejects.toThrow(
+        'Invalid server token; caused by AuthenticationError: Server-to-server token had no exp claim',
       );
     });
   });