+++ plugins/pagerduty/src/components/ChangeEvents/ChangeEvents.test.tsx
@@ -15,7 +15,7 @@
  */
 import React from 'react';
 import { render, waitFor } from '@testing-library/react';
-import { ChangeEvent } from '../types';
+import { PagerDutyChangeEvent } from '../types';
 import { TestApiRegistry, wrapInTestApp } from '@backstage/test-utils';
 import { pagerDutyApiRef } from '../../api';
 import { ApiProvider } from '@backstage/core-app-api';
@@ -30,7 +30,7 @@ describe('Incidents', () => {
   it('Renders an empty state when there are no change events', async () => {
     mockPagerDutyApi.getChangeEventsByServiceId = jest
       .fn()
-      .mockImplementationOnce(async () => []);
+      .mockImplementationOnce(async () => ({ change_events: [] }));
 
     const { getByText, queryByTestId } = render(
       wrapInTestApp(
@@ -46,37 +46,36 @@ describe('Incidents', () => {
   it('Renders all change events', async () => {
     mockPagerDutyApi.getChangeEventsByServiceId = jest
       .fn()
-      .mockImplementationOnce(
-        async () =>
-          [
-            {
-              id: 'id1',
-              source: 'changeSource1',
-              html_url: 'www.pdlink.com',
-              links: [
-                {
-                  href: 'www.externalLink1.com',
-                  text: 'link1',
-                },
-              ],
-              summary: 'summary of event',
-              timestamp: '2020-07-17T08:42:58.315+0000',
-            },
-            {
-              id: 'id2',
-              source: 'changeSource1',
-              html_url: 'www.pdlink.com/link',
-              links: [
-                {
-                  href: 'www.externalLink1.com',
-                  text: 'link1',
-                },
-              ],
-              summary: 'sum of EVENT',
-              timestamp: '2020-07-18T08:42:58.315+0000',
-            },
-          ] as ChangeEvent[],
-      );
+      .mockImplementationOnce(async () => ({
+        change_events: [
+          {
+            id: 'id1',
+            source: 'changeSource1',
+            html_url: 'www.pdlink.com',
+            links: [
+              {
+                href: 'www.externalLink1.com',
+                text: 'link1',
+              },
+            ],
+            summary: 'summary of event',
+            timestamp: '2020-07-17T08:42:58.315+0000',
+          },
+          {
+            id: 'id2',
+            source: 'changeSource1',
+            html_url: 'www.pdlink.com/link',
+            links: [
+              {
+                href: 'www.externalLink1.com',
+                text: 'link1',
+              },
+            ],
+            summary: 'sum of EVENT',
+            timestamp: '2020-07-18T08:42:58.315+0000',
+          },
+        ] as PagerDutyChangeEvent[],
+      }));
     const { getByText, getAllByTitle, queryByTestId } = render(
       wrapInTestApp(
         <ApiProvider apis={apis}>
@@ -95,36 +94,35 @@ describe('Incidents', () => {
   it('Does not render a pagerduty link when html_url is not present in response', async () => {
     mockPagerDutyApi.getChangeEventsByServiceId = jest
       .fn()
-      .mockImplementationOnce(
-        async () =>
-          [
-            {
-              id: 'id1',
-              source: 'changeSource1',
-              links: [
-                {
-                  href: 'www.externalLink1.com',
-                  text: 'link1',
-                },
-              ],
-              summary: 'summary of event',
-              timestamp: '2020-07-17T08:42:58.315+0000',
-            },
-            {
-              id: 'id2',
-              source: 'changeSource1',
-              html_url: 'www.pdlink.com/link',
-              links: [
-                {
-                  href: 'www.externalLink1.com',
-                  text: 'link1',
-                },
-              ],
-              summary: 'sum of EVENT',
-              timestamp: '2020-07-18T08:42:58.315+0000',
-            },
-          ] as ChangeEvent[],
-      );
+      .mockImplementationOnce(async () => ({
+        change_events: [
+          {
+            id: 'id1',
+            source: 'changeSource1',
+            links: [
+              {
+                href: 'www.externalLink1.com',
+                text: 'link1',
+              },
+            ],
+            summary: 'summary of event',
+            timestamp: '2020-07-17T08:42:58.315+0000',
+          },
+          {
+            id: 'id2',
+            source: 'changeSource1',
+            html_url: 'www.pdlink.com/link',
+            links: [
+              {
+                href: 'www.externalLink1.com',
+                text: 'link1',
+              },
+            ],
+            summary: 'sum of EVENT',
+            timestamp: '2020-07-18T08:42:58.315+0000',
+          },
+        ] as PagerDutyChangeEvent[],
+      }));
     const { getByText, getAllByTitle, queryByTestId } = render(
       wrapInTestApp(
         <ApiProvider apis={apis}>