+++ plugins/search-backend-module-elasticsearch/src/index.ts
@@ -22,10 +22,20 @@
 
 export { ElasticSearchSearchEngine } from './engines';
 export type {
+  ElasticSearchAgentOptions,
+  ElasticSearchConcreteQuery,
   ElasticSearchClientOptions,
   ElasticSearchHighlightConfig,
   ElasticSearchHighlightOptions,
+  ElasticSearchQueryTranslator,
   ElasticSearchQueryTranslatorOptions,
+  ElasticSearchConnectionConstructor,
+  ElasticSearchTransportConstructor,
+  ElasticSearchNodeOptions,
+  ElasticSearchOptions,
+  ElasticSearchAuth,
   ElasticSearchSearchEngineIndexer,
   ElasticSearchSearchEngineIndexerOptions,
+  ElasticSearchCustomIndexTemplate,
+  ElasticSearchCustomIndexTemplateBody,
 } from './engines';