+++ plugins/bazaar-backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-bazaar-backend
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.0
+  - @backstage/backend-test-utils@0.1.25
+
+## 0.1.17-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.0-next.2
+  - @backstage/backend-test-utils@0.1.25-next.2
+
 ## 0.1.17-next.0
 
 ### Patch Changes