+++ plugins/example-todo-list-backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @internal/plugin-todo-list-backend
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.0
+  - @backstage/plugin-auth-node@0.2.2
+
+## 1.0.2-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.0-next.2
+  - @backstage/plugin-auth-node@0.2.2-next.2
+
 ## 1.0.2-next.0
 
 ### Patch Changes