+++ plugins/home/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-home
 
+## 0.4.22
+
+### Patch Changes
+
+- 8f7b1835df: Updated dependency `msw` to `^0.41.0`.
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.1.1
+  - @backstage/core-components@0.9.5
+  - @backstage/core-plugin-api@1.0.3
+  - @backstage/plugin-stack-overflow@0.1.2
+  - @backstage/catalog-model@1.0.3
+
 ## 0.4.22-next.1
 
 ### Patch Changes