+++ plugins/permission-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-permission-react
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.0.3
+  - @backstage/plugin-permission-common@0.6.2
+
 ## 0.4.2-next.0
 
 ### Patch Changes