+++ plugins/rollbar-backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-rollbar-backend
 
+## 0.1.30
+
+### Patch Changes
+
+- 8f7b1835df: Updated dependency `msw` to `^0.41.0`.
+- Updated dependencies
+  - @backstage/backend-common@0.14.0
+
+## 0.1.30-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.0-next.2
+
 ## 0.1.30-next.1
 
 ### Patch Changes