+++ plugins/scaffolder-backend-module-rails/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/plugin-scaffolder-backend-module-rails
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-backend@1.3.0
+  - @backstage/backend-common@0.14.0
+  - @backstage/integration@1.2.1
+
+## 0.4.1-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-backend@1.3.0-next.2
+  - @backstage/backend-common@0.14.0-next.2
+  - @backstage/integration@1.2.1-next.2
+
 ## 0.4.1-next.0
 
 ### Patch Changes